Add extremely basic irc bot that can transclude.
There's a lot missing here, and it is very much proof of concept
code, but there's enough there to make it possible to play around
a little bit.

The next bit would be a WSGI app that can present data appropriately.
IRC logs can sort of be regenerated from the content of the database
but there needs to be some way of indicating context (the channel)
while not disrupting the idea that any nid ought to be usable in any
transcluding place.