Add a simple web app for IRC log presentation
The logs are created dynamically from the database contents.
Two URL routes are available:

* /<guid> redirects to that piece of content within a log showing up
  to an hour of content on either side of this content. The redirect
  includes a #fragment to scroll to the right place.

* /logs/<context>?dated=<iso format date time>. This displays the log
  described above. If there is no content at this time, search backwards
  up to twelve hours. Otherwise return no content.

A context is based on the url in the database, in this case the name of
an IRC channel.

The HTML output goes through a jinja2 template.

There's a lot wrong with this. A great deal of configuration is static
in the code. A lot of the features are too specific to IRC.

Some types of content should not display a log, instead it should just
redrect to that content in its static form (for example a mailing list)
but with a #<guid> on the URL. These sorts of situations will use url
field in the database properly.

But for now want to get the irc part working.