Make db url configurable over mod_wsgi, add some irc commands
Lots of static settings here. Probably makes sense to have a config
file, but one that works with argparse (apparently this is possible).
Or to do more dynamic determination of the things that are being
configured. In this change it was important to get the features
working (this now runs against a real wsgi server with mysql) than
to have the code correct.

Things like: where do templates lives, what is the db url, what is
the url of the server where logs live.

The irc command sare p!logs p!log and p!help. The first two show
links, the third explains the first two. p! was used a prefix to be
more unique than just !.

At some point it will be important to chase concerns about
concurrency.