Correct connection loss issues with mysql add p!hist
If the mysql connections are idle for a long time, they will be
cleaned up. So, borrowring from tiddlywebplugins.mysql we now
ping on connect and if the database is not there cause a new
connection.

The p!hist command gives up to the ten last messages in the channel.

Considering a p!spy to get the ten last uses of your nickname in the
channel.