Enable the bot to listen to private messages for commands
For this to work, some of the commands need to be passed the
channel as an argument. Therefore anything after the command
passed to a parser which spits on the first space and tries
extract and channel and a number. All five of these should work
from a public message. Only those with a channel will work in a
private message:

p!spy
p!spy 5
p!spy somechannel
p!spy 5 somechannel
p!spy somechannel 5

In the process the database tables have been tuned up so they work
better for both msyql and sqlite.