Adjust CSS and irc template HTML for improved presentation
This is based on a suggestion from @FND. It doesn't exactly
replicate what he provided because it needs to deal with nids
and there were also some issues with wide irc nicks. There's
still plenty of room for improvement, but this is a move in
the right direction.

Also included here is improved presentation of datetime
information, by using strftime directly.